insert_medical_record/requirements.txt
functions-framework==3.*
flask
psycopg2-binary
